+++ mods/src/multiplayer/raknet.cpp
@@ -19,7 +19,7 @@ static RakNetInstance *RakNetInstance_injection(RakNetInstance_constructor_t ori
 
 // Fix RakNet::RakString Security Bug
 //
-// RakNet::RakString's format constructor is often given un-sanitized user input and is never used for formatting,
+// RakNet::RakString's format constructor is often given unsanitized user input and is never used for formatting;
 // this is a massive security risk, allowing clients to run arbitrary format specifiers, this disables the
 // formatting functionality.
 typedef RakNet_RakString *(*RakNet_RakString_constructor_2_t)(RakNet_RakString *self, const char *format, ...);
@@ -56,6 +56,17 @@ static RakNet_StartupResult RakNetInstance_host_RakNet_RakPeer_Startup_injection
     return result;
 }
 
+// Properly Deallocate Pending Players
+static void ServerSideNetworkHandler_onDisconnect_injection(ServerSideNetworkHandler_onDisconnect_t original, ServerSideNetworkHandler *self, const RakNet_RakNetGUID &guid) {
+    // Call Original Method
+    original(self, guid);
+    // Free
+    Player *player = self->popPendingPlayer(guid);
+    if (player) {
+        player->destructor_deleting();
+    }
+}
+
 // Init
 void _init_multiplayer_raknet() {
     // Fix Bug Where RakNetInstance Starts Pinging Potential Servers Before The "Join Game" Screen Is Opened
@@ -71,5 +82,10 @@ void _init_multiplayer_raknet() {
     // Print Error Message If RakNet Startup Fails
     if (feature_has("Log RakNet Startup Errors", server_enabled)) {
         overwrite_call((void *) 0x73778, RakNet_RakPeer_Startup, RakNetInstance_host_RakNet_RakPeer_Startup_injection);
+    }
+
+    // Free Pending Players
+    if (feature_has("Properly Free Pending Players", server_enabled)) {
+        overwrite_calls(ServerSideNetworkHandler_onDisconnect, ServerSideNetworkHandler_onDisconnect_injection);
     }
 }