+++ docs/CHANGELOG.md
@@ -94,6 +94,7 @@ include_toc: true
     * `Fix Torch Placement`
     * `Fix Eggs Spawning Abnormally Healthy Chickens`
     * `Fix Tile Entity Syncing`
+    * `Properly Free Pending Players`
   * Existing Functionality (All Enabled By Default)
     * `Fix Screen Rendering When Hiding HUD`
     * `Sanitize Usernames`